transfer httpStattusCode to constants; rename custom response error
