+++ server/src/middleware/errorHandler.middleware.ts
@@ -1,7 +1,7 @@
 import { Request, Response, NextFunction } from 'express';
-import { ErrorHandler, handleError } from '../helpers/errorHandler.helper/errorHandler.helper';
+import { ErrorResponse, handleError } from '../helpers/errorHandler.helper';
 
-const errorHandlerMiddleware = (err: ErrorHandler, req: Request, res: Response, next: NextFunction) => {
+const errorHandlerMiddleware = (err: ErrorResponse, req: Request, res: Response, next: NextFunction) => {
 	handleError(err, res);
 };
